feat: レイアウトを AppLayout に分離 - 共通メニュー実装
